Refactor move custom ModalDialog to its own file
Refactor rename ModalDialog to SDModalDialog

This dialog encapsulates SecureDrop-specific styles and behavior,
and is used as the basis for most dialogs in the application.

Note: I'm suffering a bit with the combination of namespace
flattening and mock-heavy tests.
