+++ securedrop_client/gui/__init__.py
@@ -1,5 +1,8 @@
 """
-Generic custom widgets.
+Graphical user interface (GUI) elements.
+
+This package contains mostly Qt widgets, some of them specific
+to SecuredDrop.
 
 Copyright (C) 2018  The Freedom of the Press Foundation.
 
@@ -19,6 +22,7 @@ along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
 # Expose GUI elements without exposing the package structure
 from securedrop_client.gui.buttons import SDPushButton  # noqa: F401
+from securedrop_client.gui.dialogs import SDModalDialog  # noqa: F401
 from securedrop_client.gui.misc import (  # noqa: F401
     SecureQLabel,
     SvgLabel,
