+++ securedrop_client/gui/widgets.py
@@ -33,19 +33,15 @@ from PyQt5.QtGui import (
     QFocusEvent,
     QFont,
     QIcon,
-    QKeyEvent,
     QKeySequence,
     QLinearGradient,
     QMouseEvent,
     QPalette,
-    QPixmap,
     QResizeEvent,
 )
 from PyQt5.QtWidgets import (
     QAction,
-    QApplication,
     QDialog,
-    QDialogButtonBox,
     QGraphicsDropShadowEffect,
     QGridLayout,
     QHBoxLayout,
@@ -76,6 +72,7 @@ from securedrop_client.db import (
 )
 from securedrop_client.export import ExportError, ExportStatus
 from securedrop_client.gui import (
+    SDModalDialog,
     SDPushButton,
     SecureQLabel,
     SvgLabel,
@@ -2401,186 +2398,7 @@ class FileWidget(QWidget):
         self._set_file_state()
 
 
-class ModalDialog(QDialog):
-
-    BUTTON_CSS = load_css("modal_dialog_button.css")
-    ERROR_DETAILS_CSS = load_css("modal_dialog_error_details.css")
-
-    MARGIN = 40
-    NO_MARGIN = 0
-
-    def __init__(self, show_header: bool = True, dangerous: bool = False) -> None:
-        parent = QApplication.activeWindow()
-        super().__init__(parent)
-        self.setObjectName("ModalDialog")
-        self.setModal(True)
-
-        self.show_header = show_header
-        self.dangerous = dangerous
-        if self.dangerous:
-            self.setProperty("class", "dangerous")
-
-        # Widget for displaying error messages
-        self.error_details = QLabel()
-        self.error_details.setObjectName("ModalDialog_error_details")
-        self.error_details.setStyleSheet(self.ERROR_DETAILS_CSS)
-        self.error_details.setWordWrap(True)
-        self.error_details.hide()
-
-        # Body to display instructions and forms
-        self.body = QLabel()
-        self.body.setObjectName("ModalDialog_body")
-        self.body.setWordWrap(True)
-        self.body.setScaledContents(True)
-        body_container = QWidget()
-        self.body_layout = QVBoxLayout()
-        self.body_layout.setContentsMargins(
-            self.NO_MARGIN, self.NO_MARGIN, self.NO_MARGIN, self.NO_MARGIN
-        )
-        body_container.setLayout(self.body_layout)
-        self.body_layout.addWidget(self.body)
-
-        # Main widget layout
-        layout = QVBoxLayout(self)
-        layout.setContentsMargins(self.MARGIN, self.MARGIN, self.MARGIN, self.MARGIN)
-        self.setLayout(layout)
-
-        if self.show_header:
-            # Header for icon and task title
-            header_container = QWidget()
-            header_container_layout = QHBoxLayout()
-            header_container.setLayout(header_container_layout)
-            self.header_icon = SvgLabel("blank.svg", svg_size=QSize(64, 64))
-            self.header_icon.setObjectName("ModalDialog_header_icon")
-            self.header_spinner = QPixmap()
-            self.header_spinner_label = QLabel()
-            self.header_spinner_label.setObjectName("ModalDialog_header_spinner")
-            self.header_spinner_label.setMinimumSize(64, 64)
-            self.header_spinner_label.setVisible(False)
-            self.header_spinner_label.setPixmap(self.header_spinner)
-            self.header = QLabel()
-            self.header.setObjectName("ModalDialog_header")
-            header_container_layout.addWidget(self.header_icon)
-            header_container_layout.addWidget(self.header_spinner_label)
-            header_container_layout.addWidget(self.header, alignment=Qt.AlignCenter)
-            header_container_layout.addStretch()
-
-            self.header_line = QWidget()
-            self.header_line.setObjectName("ModalDialog_header_line")
-
-            layout.addWidget(header_container)
-            layout.addWidget(self.header_line)
-
-        layout.addWidget(self.error_details)
-        layout.addWidget(body_container)
-        layout.addWidget(self.configure_buttons())
-
-        # Activestate animation.
-        self.button_animation = load_movie("activestate-wide.gif")
-        self.button_animation.setScaledSize(QSize(32, 32))
-        self.button_animation.frameChanged.connect(self.animate_activestate)
-
-        # Header animation.
-        self.header_animation = load_movie("header_animation.gif")
-        self.header_animation.setScaledSize(QSize(64, 64))
-        self.header_animation.frameChanged.connect(self.animate_header)
-
-    def configure_buttons(self) -> QWidget:
-        # Buttons to continue and cancel
-        window_buttons = QWidget()
-        window_buttons.setObjectName("ModalDialog_window_buttons")
-
-        button_layout = QVBoxLayout()
-        window_buttons.setLayout(button_layout)
-
-        self.cancel_button = QPushButton(_("CANCEL"))
-        self.cancel_button.setStyleSheet(self.BUTTON_CSS)
-        self.cancel_button.clicked.connect(self.close)
-
-        self.continue_button = QPushButton(_("CONTINUE"))
-        self.continue_button.setStyleSheet(self.BUTTON_CSS)
-        self.continue_button.setIconSize(QSize(21, 21))
-
-        button_box = QDialogButtonBox(Qt.Horizontal)
-        button_box.setObjectName("ModalDialog_button_box")
-
-        if self.dangerous:
-            self.cancel_button.setAutoDefault(True)
-            self.continue_button.setDefault(False)
-            self.cancel_button.setObjectName("ModalDialog_primary_button")
-            self.continue_button.setObjectName("ModalDialog_cancel_button")
-        else:
-            self.cancel_button.setAutoDefault(False)
-            self.continue_button.setDefault(True)
-            self.cancel_button.setObjectName("ModalDialog_cancel_button")
-            self.continue_button.setObjectName("ModalDialog_primary_button")
-
-        button_box.addButton(self.cancel_button, QDialogButtonBox.ActionRole)
-        button_box.addButton(self.continue_button, QDialogButtonBox.ActionRole)
-
-        self.confirmation_label = QLabel()
-        self.confirmation_label.setObjectName("ModalDialogConfirmation")
-        button_layout.addWidget(self.confirmation_label, 0, Qt.AlignLeft | Qt.AlignBottom)
-
-        button_layout.addWidget(button_box, alignment=Qt.AlignLeft)
-
-        button_layout.setContentsMargins(
-            self.NO_MARGIN, self.NO_MARGIN, self.NO_MARGIN, self.NO_MARGIN
-        )
-
-        return window_buttons
-
-    def keyPressEvent(self, event: QKeyEvent) -> None:
-        if event.key() == Qt.Key_Enter or event.key() == Qt.Key_Return:
-            if self.cancel_button.hasFocus():
-                self.cancel_button.click()
-            else:
-                self.continue_button.click()
-        else:
-            super().keyPressEvent(event)
-
-    def animate_activestate(self) -> None:
-        self.continue_button.setIcon(QIcon(self.button_animation.currentPixmap()))
-
-    def animate_header(self) -> None:
-        self.header_spinner_label.setPixmap(self.header_animation.currentPixmap())
-
-    def start_animate_activestate(self) -> None:
-        self.button_animation.start()
-        self.continue_button.setText("")
-        self.continue_button.setMinimumSize(QSize(142, 43))
-        # Reset widget stylesheets
-        self.continue_button.setStyleSheet("")
-        self.continue_button.setObjectName("ModalDialog_primary_button_active")
-        self.continue_button.setStyleSheet(self.BUTTON_CSS)
-        self.error_details.setStyleSheet("")
-        self.error_details.setObjectName("ModalDialog_error_details_active")
-        self.error_details.setStyleSheet(self.ERROR_DETAILS_CSS)
-
-    def start_animate_header(self) -> None:
-        self.header_icon.setVisible(False)
-        self.header_spinner_label.setVisible(True)
-        self.header_animation.start()
-
-    def stop_animate_activestate(self) -> None:
-        self.continue_button.setIcon(QIcon())
-        self.button_animation.stop()
-        self.continue_button.setText(_("CONTINUE"))
-        # Reset widget stylesheets
-        self.continue_button.setStyleSheet("")
-        self.continue_button.setObjectName("ModalDialog_primary_button")
-        self.continue_button.setStyleSheet(self.BUTTON_CSS)
-        self.error_details.setStyleSheet("")
-        self.error_details.setObjectName("ModalDialog_error_details")
-        self.error_details.setStyleSheet(self.ERROR_DETAILS_CSS)
-
-    def stop_animate_header(self) -> None:
-        self.header_icon.setVisible(True)
-        self.header_spinner_label.setVisible(False)
-        self.header_animation.stop()
-
-
-class PrintDialog(ModalDialog):
+class PrintDialog(SDModalDialog):
 
     FILENAME_WIDTH_PX = 260
 
@@ -2707,7 +2525,7 @@ class PrintDialog(ModalDialog):
                 self._show_generic_error_message()
 
 
-class ExportDialog(ModalDialog):
+class ExportDialog(SDModalDialog):
 
     PASSPHRASE_LABEL_SPACING = 0.5
     NO_MARGIN = 0
@@ -2961,7 +2779,7 @@ class ExportDialog(ModalDialog):
                 self._show_generic_error_message()
 
 
-class DeleteSourceDialog(ModalDialog):
+class DeleteSourceDialog(SDModalDialog):
     """Used to confirm deletion of source accounts."""
 
     def __init__(self, source: Source, controller: Controller) -> None:
@@ -3009,7 +2827,7 @@ class DeleteSourceDialog(ModalDialog):
         self.close()
 
 
-class DeleteConversationDialog(ModalDialog):
+class DeleteConversationDialog(SDModalDialog):
     """
     Shown to confirm deletion of all content in a source conversation.
     """
